Getting started refactoring hydrate
